requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
commonmark==0.9.1
defusedxml==0.7.1
Django==3.2.16
django-cors-headers==3.13.0
django-currentuser==0.5.3
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
httpie==3.2.1
idna==3.4
multidict==6.0.4
Pillow==9.4.0
Pygments==2.14.0
PyJWT==1.7.1
PySocks==1.7.1
pytz==2022.7.1
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.1.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
